namespace refactoring complete
