+++ config/hyn.php
@@ -1,18 +1,22 @@
 <?php
 
+use Hyn\ManagementInterface\ManagementInterfaceServiceProvider;
+use Hyn\Tenancy\TenancyServiceProvider;
+use Hyn\Webserver\WebserverServiceProvider;
+
 return [
     'packages' => [
         'multi-tenant'         => [
             'description'      => 'Multi tenancy for Laravel 5',
-            'service-provider' => 'Hyn\Tenancy\MultiTenantServiceProvider',
+            'service-provider' => TenancyServiceProvider::class,
         ],
         'management-interface' => [
             'description'      => 'Interface for managing webserver and multi tenancy',
-            'service-provider' => 'Hyn\ManagementInterface\ManagementInterfaceServiceProvider',
+            'service-provider' => ManagementInterfaceServiceProvider::class,
         ],
         'webserver'            => [
             'description'      => 'Integration into and generation of configs for webservices',
-            'service-provider' => 'Hyn\Webserver\WebserverServiceProvider',
+            'service-provider' => WebserverServiceProvider::class,
         ],
     ],
 ];
