+++ src/Webserver/WebserverServiceProvider.php
@@ -23,13 +23,24 @@ class WebserverServiceProvider extends ServiceProvider
     {
 
         // configuration
-        $this->mergeConfigFrom(__DIR__.'/../../config/webserver.php', 'webserver');
-        $this->publishes([__DIR__ . '/../../config/webserver.php' => config_path('webserver.php')],
-            'hyn-webserver-config');
+        $this->mergeConfigFrom(
+            __DIR__ . '/../../config/webserver.php',
+            'webserver'
+        );
+        $this->publishes(
+            [__DIR__ . '/../../config/webserver.php' => config_path('webserver.php')],
+            'hyn-webserver-config'
+        );
+
         // adds views
-        $this->loadViewsFrom(__DIR__ . '/../../views/webserver', 'webserver');
+        $this->loadViewsFrom(
+            __DIR__ . '/../../views/webserver',
+            'webserver'
+        );
+
         // migrations
-        $this->publishes([__DIR__ . '/../../database/migrations/' => database_path('/migrations')],
+        $this->publishes(
+            [__DIR__ . '/../../database/migrations/' => database_path('/migrations')],
             'hyn-webserver-migrations');
 
         Website::observe(new Observers\WebsiteObserver());
